simpler TUI (#6588)
* initial new ID-centric TUI

plus an unrelated change (sorry): cpuprofile and pprof are now env vars
instead of flags, since I noticed the flag parsing didn't seem to
consistently, and we hide them anyway

as part of this, ID.Digest now caches

Signed-off-by: Alex Suraci <alex@dagger.io>

* optimize

Signed-off-by: Alex Suraci <alex@dagger.io>

* clean up error handling for failed commands

this could probably be split up:

* fix repeated error message
* return a generic error when getting the ID fails during JSON
  marshaling, rather than bubbling up the ID error, since it should
  already be represented elsewhere in the progress stream
* display combined stdout/stderr stream instead of split, which is noisy

Signed-off-by: Alex Suraci <alex@dagger.io>

* idtui: show task progress, simplify durations, no spinners

Signed-off-by: Alex Suraci <alex@dagger.io>

* refactor interestingness, add Debug

Signed-off-by: Alex Suraci <alex@dagger.io>

* idtui: support terminals

TODO:

* render a status bar up top so you can switch back to the original
  output (like a tab)
* scrollable viewport
* print all logs on exit, instead of just visible viewport?

Signed-off-by: Alex Suraci <alex@dagger.io>

* better step display heuristics

goals:

* avoid flickering by only removing steps after 1 second
* only show steps that took >100ms to avoid an overwhelming amount of
  noise

justification:

This new TUI should converge to only showing things that are actively
running. For example, 'dagger up' output should be nice and clean, only
showing the tunnels that were created and the services that are running.

We don't want the user to have to hand-configure hints for what to show,
and we don't want to accidentally hide critical things like "stuck"
steps. So we'll treat the duration of the step itself as the heuristic.

In the case of `dagger call`, the value you're querying will already be
printed on exit. Any other output is extra; the only reason you would
care about it is when it's taking too long. So, we only show what's
taking long.

Signed-off-by: Alex Suraci <alex@dagger.io>

* shorter initial polling interval for connect

this is called a lot of times in a pipeline, might as well shave off
what we can

Signed-off-by: Alex Suraci <alex@dagger.io>

* fix flicker when exiting zoomed terminal

+ fix showing 'canceled' instead of final paint
+ general cleanup

Signed-off-by: Alex Suraci <alex@dagger.io>

* set log terminal width

otherwise it auto-resizes indefinitely which tanks perf with a lot of
output

Signed-off-by: Alex Suraci <alex@dagger.io>

* dagql/idproto: rename Parent to Base

Right now we have Parent meaning two things:

1. the ancestor of an ID, i.e. the call chain leading up to it
1. the caller of another set of vertices, otel parent span style

This renames the ID "Parent" to "Base" since it's kinda similar to base
layers in containers.

Signed-off-by: Alex Suraci <alex@dagger.io>

* ID modules now include name + ref

...plus a lot of UI experiments (sorry again)

Signed-off-by: Alex Suraci <alex@dagger.io>

* restore yellow tickers and green checks

Signed-off-by: Alex Suraci <alex@dagger.io>

* assimilate more responsibility into Frontend

* Use a Vertex with a well-known ID (idtui.PrimaryVertex) to record the
  primary output, since we want these to also show up in Cloud
* Frontend is the authoritative handler for writing to os.Stdout and
  os.Stderr; otherwise you can get jumbled output when piping to things.
  * For example, dagger query used to write directly to
    os.Stdout/os.Stderr if it saw they weren't a TTY. Now it writes to
    the primary vertex, where it's held in a buffer until we stop the
    TUI and do the final print.
* Removed a few Vertexes and Tasks around module loading; they don't
  really help clarify things in the new TUI, and in the long run we want
  these to be engine-side function calls instead
* Frontend now handles Plain and Silent modes too, otherwise we end up
  with tricky conditionals that lead to bad behavior in subtle places
* For backwards compatibility, dagger run still uses the old TUI.

Signed-off-by: Alex Suraci <alex@dagger.io>

* support plain output for dagger run/legacy TUI

Signed-off-by: Alex Suraci <alex@dagger.io>

* undo CombinedOutput

still want this, but it explodes the scope too large (SDK + docs changes
needed)

Signed-off-by: Alex Suraci <alex@dagger.io>

* appease linter

Signed-off-by: Alex Suraci <alex@dagger.io>

* roll back error silencing change

a test depends on this, needs more thought, unsurprisingly

Signed-off-by: Alex Suraci <alex@dagger.io>

* fix dagger up test

We need the TUI now that we only print to the ID vertex, which is marked
internal to not overwhelm the legacy TUI for dagger run.

Signed-off-by: Alex Suraci <alex@dagger.io>

* display progrock messages

same technique as before, except we print all of them on the final
render

Signed-off-by: Alex Suraci <alex@dagger.io>

* remove extraneous silence check

this is handled in the Frontend now

Signed-off-by: Alex Suraci <alex@dagger.io>

* add message to dagger init, fix dagger config

Signed-off-by: Alex Suraci <alex@dagger.io>

* avoid stack overflow while processing steps

this one's weird. it's not an obvious infinite recursion, but it seems
to be high enough to trip Go's stack limit. the problem goes away when
adding a runtime.Gosched() or log.Println in HighLevelStep.

turns out we don't even need to recurse, so now we just store a step's
parent digest on it, instead of its parent step.

Signed-off-by: Alex Suraci <alex@dagger.io>

* terminal exits with error on failure

Signed-off-by: Alex Suraci <alex@dagger.io>

* idtui: fix hanging/staying open when connect fails

Signed-off-by: Alex Suraci <alex@dagger.io>

* remove TODO

Signed-off-by: Alex Suraci <alex@dagger.io>

* smaller window size

Signed-off-by: Alex Suraci <alex@dagger.io>

* use cobra.OnFinalize instead of PersistentPostRun

PersistentPostRun is not called when the command returns an error

Signed-off-by: Alex Suraci <alex@dagger.io>

* use filepath.Join instead of filepath.Rel

otherwise this fails with e.g.:

  can't make go relative to /

Signed-off-by: Alex Suraci <alex@dagger.io>

* make ID digest cache opt-in

Signed-off-by: Alex Suraci <alex@dagger.io>

* disable IsTransitiveDependency

Signed-off-by: Alex Suraci <alex@dagger.io>

---------

Signed-off-by: Alex Suraci <alex@dagger.io>