fix: all paths should be encoded over-the-wire (#7118)
* fix: all paths should be encoded over-the-wire

Somehow, this was just being done entirely incorrectly, even though I
thought I got it the first time. The path conversions are now moved to
where we directly send/recv on the wire - then we only have to have a
minimal number of conversions.

Signed-off-by: Justin Chadwell <me@jedevc.com>

* ci: add option to list all lint errors

Signed-off-by: Justin Chadwell <me@jedevc.com>

* fix: wrap all errors using %w when formatting

This was causing some details of important errors to be lost! While I
don't *think* this was causing any bugs right now, I encountered this
while trying to work with propagated grpc status codes.

We can enable a linting check to make sure that this doesn't happen
again.

Signed-off-by: Justin Chadwell <me@jedevc.com>

* fix: wrap file not exists errors with codes.NotFound

Different platforms have different error messages. And of course, on
windows, you get different error messages for `Stat` and `ReadFile`,
which is *super* fun. So instead of trying to cover every single little
possible message, we should actually have the client handle this neatly.

Signed-off-by: Justin Chadwell <me@jedevc.com>

---------

Signed-off-by: Justin Chadwell <me@jedevc.com>