feat(core): allow other CIs than GitHub (#7511)
* feat: extend parseRefString heuristic to allow more CIs

Previously, local / git refs were estimated using such heuristic:
- was being considered as remote any ref with no version specified AND ref starting with "github.com"

However, as we are now extending the logic to other CI providers, this
needs to be updated.

The current logic is extended:
- stat if the dir is present
- if it is not, we then try to estimate the root of the remote (repoRootForImportPath)
- If it fails, we fallback as a local remote (same behavior as it was in the past)

The main reason to fallback as a local remote is that a lot of `dagger install` / `dagger call` design decisions have been implemented, and is not part of the scope of this PR. e.g. `dagger init --source=new_dir` if new_dir is not present

As part of this commit, we also implement splitRootAndSubdir(), which extracts the root of the repo and the subdir from the ref.

Supported refs at the moment:
- Public CIs: github, gitlab, gitea, bitbucket
- vanity URLs

- Path to private refs support (especially with GitLab)

Signed-off-by: grouville <guillaume@dagger.io>

* fix: ref-like local directory + add integration test

Adds an integration test regarding the edge case of having a local ref that looks like a remote one: github.com/dagger/dagger, in a git context

It also makes sure that moduleSourceResolveFromCaller's sourceRootRelPath always references a local dir to not break with our new implementation of parseRefString

This integration test suite will also be the foundation for our private repos handling test suite

Signed-off-by: grouville <guillaume@dagger.io>

* feat: add parseRefString test

Implements the foundation to test parseRefString without side-effects. In
order to do so, we rely on a interface to mock buildkit's stat.

Contrary to the integration tests, this test suite is used to give
insights on how parseRefString behaves on public repos, and how it returns
informations later used by TestSplitRootAndSubdir

Signed-off-by: Guillaume de Rouville <guillaume@dagger.io>

* fix: move ExtractRootAndSubdirFromRef to vcs package, refactor + change tests

As parts of ExtractRootAndSubdirFromRef's logic are going to be shared between the core and the cmd packages, we move the code to the vcs package for clarity.
Also, this commit contains the refactor of ExtractRootAndSubdirFromRef's function + a new test suite implementation

Signed-off-by: Guillaume de Rouville <guillaume@dagger.io>

* chore: refactor to remove complex ExtractRootAndSubdirFromRef

Signed-off-by: Justin Chadwell <me@jedevc.com>

* chore: make buildkitClient interface private + remove log + extend unit test

As suggested on the comments, make the buildkitClient private and remove slog to have a silent failure
It also extends the unit test to take into account Justin's improvements

Signed-off-by: Guillaume de Rouville <guillaume@dagger.io>

* chore: regenerate sdks / docs + remove remaining cloneURL

CloneURL function was still part of the moduleSchema, which led to some conflicts.

At the same time, we regenerate the docs and SDKs

Signed-off-by: Guillaume de Rouville <guillaume@dagger.io>

* fix: improve parseRefString to also trim "/" prefix in subdir

In parseRefString, trimming the extra "/" is important as subpath traversal such as /../ are
being cleaned by filePath.Clean as just "/", making possible the escape
outside the module context

Signed-off-by: Guillaume de Rouville <guillaume@dagger.io>

* chore: add release note

Adds a changie release note

Signed-off-by: Guillaume de Rouville <guillaume@dagger.io>

---------

Signed-off-by: grouville <guillaume@dagger.io>
Signed-off-by: Guillaume de Rouville <guillaume@dagger.io>
Signed-off-by: Justin Chadwell <me@jedevc.com>
Co-authored-by: grouville <guillaume@dagger.io>
Co-authored-by: Justin Chadwell <me@jedevc.com>