Dynamically serve module schemas based on engineVersion (#7759)
* chore: split typedefs query into embedded file

This query is quite useful to manually run against a dagger instance to
see what's going on - so I moved it to a file to more easily run it
using `dagger query`.

Signed-off-by: Justin Chadwell <me@jedevc.com>

* chore: bump client/engine version compat to v0.12.0

With changes to Terminal, and other related recent changes, we should
bump the client/engine compat to the next release.

Signed-off-by: Justin Chadwell <me@jedevc.com>

* chore: add dagql views

This patch introduces dagql views - a new construct for allowing
different servers to have a unique view of the same underlying schema.

When new fields are declared, they can be modified using `.View()`,
which signifies that that field should only be reachable when accessed
under a view that matches the target.

Most of this is just some pretty simple plumbing back-and-forth -
however, we also need to ensure that we add the original view to the
underlying callpbv1 proto definition. This ensures that we track the
view that a call was made in, so that if we take its ID, and begin
evaluating it in the context of *another* view, we can still perform the
desired operation in it's original view.

Signed-off-by: Justin Chadwell <me@jedevc.com>

* tests: add test for terminal in middle of pipeline

Signed-off-by: Justin Chadwell <me@jedevc.com>

* chore: add plumbing to allow custom core api views per-module

This adds support for the previously introduced dagql views into the
core API. Essentially, each view is a set of valid versions, with the
target view being the version defined in the `dagger.json`
`engineVersion` field.

This version is configured on each `dagql.Server` (very carefully,
sometimes we destructively modify it, in the case of serving a module,
and other cases non-destructively, in the case of creating copies of
dependencies).

Then we have one schema, with different views per version, so that we
can build backwards compat support for various features.

Signed-off-by: Justin Chadwell <me@jedevc.com>

* tests: add module tests for custom core views

This tests a variety of configuration scenarios, to ensure that each
module gets a correct view of the world.

To do this, we introduce a new `schemaVersion` property, which gets the
version of the schema we're currently being served, and allows for
easily testing this.

Signed-off-by: Justin Chadwell <me@jedevc.com>

* feat: add backwards compat for export boolean

This adds backwards compat for before v0.12.0 for dagger/dagger#7500,
which changed the return type for export from a boolean to a string.
This shouldn't be a breaking change for many people - but, 1. it might
for some, and we can easily avoid it, and 2. it's a nice easy test for
our new module version compat.

Signed-off-by: Justin Chadwell <me@jedevc.com>

* feat: add backwards compat for new terminal

This adds backwards compat for before v0.12.0 for dagger/dagger#7586,
which changed the Container.terminal field to return a Container type
instead of a terminal. This is a breaking change for users who were
manually calling Terminal in their code, as well as accepting/returning
it as an argument.

This patch restores the old behavior for before v0.12.0, and also
attempts to insert a shim the old behavior, so we get consistent
results.

Signed-off-by: Justin Chadwell <me@jedevc.com>

* fix: don't insert global view into ids

Signed-off-by: Justin Chadwell <me@jedevc.com>

* feat: dagger develop should update engine version before running codegen

Since now the engineVersion field is now used to influence codegen, the
ordering now matters.

Additionally, to control this with more detail, we have
`--engine-version` as a parameter to `dagger develop` that can be used
to set a specific dagger schema version to target (was helpful during
manual testing), and can also be set to the empty string to skip
updating the engine version from what it was before.

Signed-off-by: Justin Chadwell <me@jedevc.com>

* chore: add changelog

Signed-off-by: Justin Chadwell <me@jedevc.com>

* ci: add engine generation+linting for protobuf

Signed-off-by: Justin Chadwell <me@jedevc.com>

* chore: rename schemaVersion to __schemaVersion

Signed-off-by: Justin Chadwell <me@jedevc.com>

---------

Signed-off-by: Justin Chadwell <me@jedevc.com>