feat: add `Directory.asGit` (#9730)
* chore: move matchVersion helpers to modulesource.go

These are exclusively used in modulesource.go, so we should move them to
there.

Signed-off-by: Justin Chadwell <me@jedevc.com>

* tests: refactor GitTest to avoid testutil.Query

Signed-off-by: Justin Chadwell <me@jedevc.com>

* feat: add Directory.asGit

Signed-off-by: Justin Chadwell <me@jedevc.com>

* feat: refactor dagop usage

This allows returning `File`s and `Directory`s more easily, using a
wrapped immutable cache ref - this avoids needing our nasty hack to
import from the host file system: we can just write directly to our
cache ref.

To get this working with `Container.asTarball` though, we needed a
little reworking, so that's it's possible to get a reasonable filename
for the tarball. Potentially this doesn't work if the filename is
*completely* unknown until running the function? So, the
`DirectoryDagOp` becomes a generic `FSDagOp` able to handle both
`Directory`s and `File`s.

Signed-off-by: Justin Chadwell <me@jedevc.com>

* chore: suggestions from review

Signed-off-by: Justin Chadwell <me@jedevc.com>

* add support for result release callbacks, use on dir/file

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

* fix sdk call cache hits on repos w/ different auth

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

* fix: extractLLBDependencies with nil data

Signed-off-by: Justin Chadwell <me@jedevc.com>

---------

Signed-off-by: Justin Chadwell <me@jedevc.com>
Signed-off-by: Erik Sipsma <erik@sipsma.dev>
Co-authored-by: Erik Sipsma <erik@sipsma.dev>